pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-auth>=2.40.3",
    "google-cloud-aiplatform>=1.70.0,<1.85.0",
    "google-cloud-documentai>=3.6.0",
    "google-cloud-firestore>=2.21.0",
    "google-cloud-storage>=1.32.0,<3.0.0",
    "google-genai>=1.32.0",
    "pandas>=2.3.2",
    "pdf2image>=1.17.0",
    "pillow>=11.3.0",
    "plotly>=6.3.0",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.0.0",
    "streamlit>=1.49.1",
    "vertexai>=1.70.0",
]
